==> scripts/04_analysis.Rmd <==
---
title: "04_analysis"
author: "Evan Muise"
date: "5/20/2021"
output: html_document
---

Setup!

```{r setup}
library(tidyverse) # in use
library(tidymodels) # t tests
library(cowplot) # plot additions
library(patchwork) # plot additions
library(ggrepel) # for t-test plot
library(ggridges) # for ridge plot
library(magick) # for image cropping of 3d vectors
#library(gtsummary)
#library(ggpubr)
library(flextable)

library(sf) # in use
library(raster) # in use for data

library(ggsci) # in use for aaas
library(scico) # in use for ridge plot
#library(ggthemes) 

my_theme <- theme_bw() +
  theme(panel.grid = element_blank())
theme_set(my_theme)
set.seed(69420)

source(here::here("scripts", "get_keys.R"))

elev_agg <- 50
vlce_colours <- keys$vlce %>% pull(ntems_colour) %>% rev()
```

```{r load}
load(here::here("data", "backup.RData"))
```

Merged DF methods

```{r process, message=FALSE}
full_dfs <- list.files(here::here("data", "all_vars"), 
                       pattern = ".csv", full.names = T)

sample_loc <- here::here("data", "all_vars", "sampled")
dir.create(sample_loc, showWarnings = FALSE)

process_csv <- function(df) {
  
  print(df)
  
  split <- str_split(df, pattern = "/")
  if (is.character(df)) {
    df <- read_csv(df, col_types = cols())
  }

  df <- df %>%
    filter(vlce != 0) %>% # filter out unclassified pixels
    filter(elev < 32000) # filter out elevation nodata pixels (very small amount)
  
  cont <- df %>% 
    filter(vlce %in% c(230, 220, 210, 81)) %>% # forested pixels
    filter(Change_Attribution != 2) %>% # filter out harvested pixels
    dplyr::select(subzone, protected, elev:total_biomass, vlce)
  
  # for slice_sample in next pipes
  cont_cells <- cont %>% count(protected) %>% pull(n) %>% min()
  
  both <- cont %>% count(protected) %>% pull(n) %>% length()
  # check if both protected and unprotected exists
  
  if (cont_cells != 0 & both == 2) {
  
    cont <- cont %>%
      group_by(protected) %>%
      slice_sample(n = cont_cells) # get same no. pixels prot and unprot
    
    
    
    
    sample_name <- tail(split[[1]], 1)
    
    print("saving sampled df")
    write_csv(cont, here::here(sample_loc, sample_name))
    
    cont_t_tests <- cont %>%
      pivot_longer(cols = elev:total_biomass, names_to = "variable") %>%
      group_by(subzone, variable, protected) %>%
      nest() %>%
      pivot_wider(names_from = protected, values_from = data) %>%
      ungroup() %>%
      mutate(
        t_test = map2(protected, unprotected, ~{t.test(.x$value, .y$value) %>% 
                 tidy()})
      ) %>%
      dplyr::select(!c("protected", "unprotected")) %>%
      unnest(cols = t_test)
    
  } else {
    
    cont_t_tests <- tibble()
    
  }
    
  if (nrow(df) > 0) {
    
    # change wants harvest included, so operates on df, rather than disc
    change <- df %>%
      count(subzone, protected, Change_Attribution) %>%
      group_by(subzone, protected) %>%
      mutate(per_dist = n / sum(n))
    
    elev_dist <- df %>%
      group_by(subzone, protected, elev, Change_Attribution) %>%
      summarize(count = n())
    
    # remove harvested pixels post here
    
    disc <- df %>%
      filter(Change_Attribution != 2) %>% # remove harvested pixels
      dplyr::select(subzone, protected, vlce, Change_Attribution, change_year)
    
    
    vlce <- disc %>%
      count(subzone, protected, vlce) %>%
      group_by(subzone, protected) %>%
      mutate(per_cover = n / sum(n))
    
    elev_nw <- df %>% 
      filter(Change_Attribution != 2) %>% # remove harvested pixels
      left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
      filter(class_name != "water") %>%
      group_by(subzone, protected, elev, class_name) %>%
      summarize(count = n())
    
    elev_w <- df %>% 
      filter(Change_Attribution != 2) %>% # remove harvested pixels
      left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
      group_by(subzone, protected, elev, class_name) %>%
      summarize(count = n())
    
    lat_dist <- df %>%
      group_by(subzone, protected, latitude, Change_Attribution) %>%
      summarize(count = n())
    
    
  } else {
    
    disc <- tibble()
    vlce <- tibble()
    change <- tibble()
    elev_nw <- tibble()
    elev_w <- tibble()
    elev_dist <- tibble()
    lat_dist <- tibble()
    
  }
  
  outputs <- list("cont_t_tests" = cont_t_tests, 
                  "vlce" = vlce, 
                  "change" = change,
                  "elev_nw" = elev_nw,
                  "elev_w" = elev_w,
                  "elev_dist" = elev_dist,
                  "lat_dist" = lat_dist)
    
    #names(outputs) <- c("cont_t_tests", "vlce", "change")
    
  return(outputs)
  
}

outs <- map(full_dfs, process_csv)

t_tests <- bind_rows(map(outs, 1))

vlce <- bind_rows(map(outs, 2)) %>%
  ungroup() %>%
  complete(subzone, protected, vlce, 
           fill = list(n = 0, per_cover = 0))

change <- bind_rows(map(outs, 3)) %>%
  ungroup() %>%
  complete(subzone, protected, Change_Attribution, 
           fill = list(n = 0, per_dist = 0))

elev_nw <- bind_rows(map(outs, 4))

elev_w <- bind_rows(map(outs, 5))

elev_dist <- bind_rows(map(outs, 6))

lat_dist <- bind_rows(map(outs, 7))

#structure <- bind_rows(map(outs, 7))
structure <- read_csv(here::here("data", "structure_means.csv"))
```

```{r save}
save.image(file = here::here("data", "backup.RData"))
```

Elevation & Protected/Unprotected (BEC) Figures

```{r elev-bec}
elev_zone_data <- elev_nw %>% 
  ungroup() %>%
  filter(elev < 32000) %>%
  separate(subzone, into = c("zone", "subzone")) %>% 
  group_by(protected, elev, zone) %>% 
  summarize(n = sum(count)) %>% 
  mutate(perc = n / sum(n)) 

elev_20m <- elev_zone_data %>%
  mutate(elev = elev %/% elev_agg * elev_agg) %>%
  group_by(protected, elev, zone) %>%
  summarize(n = sum(n)) %>%
  mutate(perc = n / sum(n))

elev_zone_plot_20m <- elev_20m %>%
  ggplot(aes(x = elev, y = perc, fill = zone)) +
  geom_area(stat = "identity") +
  facet_wrap(~str_to_title(protected)) +
  coord_flip() +
  scale_y_continuous(labels = scales::percent) +
  labs(x = "Elevation (m)",
       y = "% Cover",
       fill = "BEC Zone") +
  theme(legend.position = "left") +
  scale_fill_igv()

protected_elev <- elev_zone_data %>%
  ungroup() %>%
  pivot_wider(values_from = n, names_from = protected) %>% 
  group_by(elev) %>%
  summarize(protected = sum(protected, na.rm = T),
            unprotected = sum(unprotected, na.rm = T)) %>%
  mutate(total = protected + unprotected,
         per_prot = protected / total,
         per_unprot = 1 - per_prot) %>%
  pivot_longer(per_prot:per_unprot, names_to = "value", values_to = "percent")

# NOTE THE GEOM SMOOTH RATHER THAN LINE, DUE TO HIGH AND LOW ELEV LOW NUMBER
# OF PIXELS
prot_elev_plot <- protected_elev %>%
  ggplot(aes(x = elev, y = per_prot)) +
  #geom_smooth() +
  #geom_point() +
  geom_line() +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  labs(x = "Elevation (m)",
       y = "% Protected") +
  coord_flip()

prot_elev_plot_20m <- protected_elev %>% 
  mutate(elev = elev %/% elev_agg * elev_agg) %>% 
  group_by(elev) %>%
  summarize(protected = sum(protected),
            unprotected = sum(unprotected)) %>%
  mutate(total = protected + unprotected,
         per_prot = protected / total,
         per_unprot = 1 - per_prot) %>%
  ggplot(aes(x = elev, y = per_prot)) +
  #geom_smooth() +
  #geom_point() +
  geom_area() +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  labs(x = NULL,
       y = "% Protected",
       fill = NULL) +
  theme(axis.text.y = element_blank(),
        legend.position = c(0.7, 0.85)) +
  coord_flip()

mosaic_elev_plot <- plot_grid(elev_zone_plot_20m, prot_elev_plot_20m,
          rel_widths = c(3, 1),
          align = 'h', axis = 'tb')

ggsave(here::here("outputs", "bec_elev.png"), plot = mosaic_elev_plot, 
       device = "png",
       height = 5.5,
       width = 8.5)

bec_legend <- cowplot::get_legend(elev_zone_plot_20m)

protected_elev %>% 
  mutate(elev = elev %/% elev_agg * elev_agg) %>% 
  group_by(elev) %>%
  summarize(protected = sum(protected),
            unprotected = sum(unprotected)) %>%
  mutate(total = protected + unprotected,
         per_prot = protected / total,
         per_unprot = 1 - per_prot,
         per_bc = total / sum(total)) %>%
  dplyr::select(elev, per_prot, per_bc) %>%
  pivot_longer(starts_with("per")) %>%
  ggplot(aes(x = elev, y = value, colour = name)) +
  #geom_smooth() +
  #geom_point() +
  geom_line() +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  scale_colour_manual(values = c("red",
                                 "blue"),
                      labels = c("% of BC",
                                 "% Protected")) +
  labs(x = "Elevation (m)",
       y = NULL,
       fill = NULL,
       colour = NULL,
       size = "Proportion of British Columbia") #+
  #theme(axis.text.y = element_blank()) +
  #coord_flip()
```

Elevation & Protected/Unprotected (VLCE) Figures

```{r elev-vlce}
elev_lcc_nw <- elev_w %>%
  filter(elev < 32000) %>%
  mutate(elev = elev %/% elev_agg * elev_agg) %>%
  group_by(protected, elev, class_name) %>%
  summarize(n = sum(count)) %>%
  mutate(perc = n / sum(n)) %>%
  left_join(keys$vlce) %>%
  mutate(name_clean = fct_rev(fct_reorder(name_clean, class_val)))

elev_lcc_20m <- elev_lcc_nw %>%
  ggplot(aes(x = elev, y = perc, fill = name_clean)) +
  geom_area(stat = "identity") +
  coord_flip() +
  facet_wrap(~str_to_title(protected)) +
  #coord_flip() +
  scale_y_continuous(labels = scales::percent) +
  scale_fill_manual(values = vlce_colours) +
  labs(x = "Elevation (m)",
       y = "% Cover",
       fill = "Land Cover") +
  theme(legend.position = "left")

mosaic_lcc_plot <- plot_grid(elev_lcc_20m, prot_elev_plot_20m,
          rel_widths = c(3, 1),
          align = 'h', axis = 'tb')

ggsave(here::here("outputs", "lcc_elev.png"), plot = mosaic_lcc_plot, device = "png")

protected_elev %>% 
  mutate(elev = elev %/% elev_agg * elev_agg) %>% 
  group_by(elev) %>%
  summarize(protected = sum(protected),
            unprotected = sum(unprotected)) %>%
  mutate(total = protected + unprotected,
         per_prot = protected / total,
         per_unprot = 1 - per_prot) %>%
  filter(unprotected == 0)

```

Elevation and Disturbance

```{r elev-dist}
elev_dist_plot <- elev_dist %>% 
  mutate(elev = elev %/% elev_agg * elev_agg,
         Change_Attribution = ifelse(!(Change_Attribution %in% c(1, 2)), 
                                     NA, 
                                     Change_Attribution)) %>%
  group_by(protected, elev, Change_Attribution) %>%
  summarize(count = sum(count)) %>%
  mutate(perc = count / sum(count),
         Change_Attribution = case_when(Change_Attribution == 1 ~ "Fire",
                                        Change_Attribution == 2 ~ "Harvest",
                                        TRUE ~ "None")) %>%
  ggplot(aes(x = elev, y = perc, fill = Change_Attribution)) +
  geom_area(stat = "identity") +
  coord_flip() +
  facet_wrap(~str_to_title(protected)) +
  #coord_flip() +
  scale_y_continuous(labels = scales::percent) +
  #scale_fill_manual(values = vlce_colours) +
  labs(x = "Elevation (m)",
       y = "% Disturbed",
       fill = "Disturbance") +
  theme(legend.position = "left")

plot_grid(elev_dist_plot, prot_elev_plot_20m,
          rel_widths = c(3, 1),
          align = 'h', axis = 'tb')
```

Donut Plots

```{r donut}
# disturbance
change_donut <- change %>%
  mutate(Change_Attribution = ifelse(!(Change_Attribution %in% c(1, 2)), 
                                     NA, 
                                     Change_Attribution)) %>%
  group_by(protected, Change_Attribution) %>%
  summarize(count = sum(n, na.rm = T)) %>%
  mutate(perc = count / sum(count),
         Change_Attribution = case_when(Change_Attribution == 1 ~ "Fire",
                                        Change_Attribution == 2 ~ "Harvest",
                                        TRUE ~ "None")) 

change_donut %>%
  ggplot(aes(x = protected, y = perc, fill = Change_Attribution)) +
  geom_bar(stat = "identity") +
  guides(fill = guide_legend(reverse = F)) +
  theme_void() +
  theme(legend.position = "bottom") +
  coord_polar(theta = "y") +
  labs(fill = "Disturbance") +
  guides(fill = guide_legend(title.position = "top",
                             title.hjust = 0.5))

# BEC
bec_donut <- vlce %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  group_by(protected, zone) %>%
  summarize(count = sum(n, na.rm = T)) %>%
  mutate(perc = count / sum(count)) 

bec_double_donut <- bec_donut %>%
  ggplot(aes(x = protected, y = perc, fill = zone)) +
  geom_bar(stat = "identity", position = "stack") +
  guides(fill = guide_legend(reverse = F)) +
  theme_void() +
  theme(legend.position = "bottom") +
  coord_polar(theta = "y") +
  labs(fill = "BEC Zone") +
  scale_fill_igv() +
  guides(fill = guide_legend(title.position = "top",
                             title.hjust = 0.5))

ggsave(here::here("outputs", "bec_double_donut.png"), 
       plot = bec_double_donut, device = "png")

bec_donut %>%
  ggplot(aes(x = zone, y = perc, fill = zone)) +
  geom_bar(stat = "identity") +
  geom_hline(yintercept = seq(0, .2, .1), lty = "dashed") +
  coord_polar() +
  scale_fill_igv() +
  facet_wrap(~str_to_title(protected)) +
  #theme_void() +
  theme(legend.position = "bottom") +
  labs(fill = NULL,
       x = NULL)

# VLCE  
vlce_donut <- vlce %>%
  group_by(protected, vlce) %>%
  summarize(count = sum(n, na.rm = T)) %>%
  left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
  mutate(perc = count / sum(count, na.rm = T),
         name_clean = fct_reorder(name_clean, vlce)) 

vlce_double_donut <- vlce_donut %>%
  ggplot(aes(x = protected, y = perc, fill = name_clean)) +
  geom_bar(stat = "identity", position = "stack") +
  guides(fill = guide_legend(reverse = F)) +
  theme_void() +
  theme(legend.position = "bottom") +
  coord_polar(theta = "y") +
  labs(fill = "Land Cover") +
  scale_fill_manual(values = rev(vlce_colours)) +
  guides(fill = guide_legend(title.position = "top",
                             title.hjust = 0.5))

ggsave(here::here("outputs", "vlce_double_donut.png"), 
       plot = vlce_double_donut, device = "png")


vlce_donut %>%
  ggplot(aes(x = name_clean, y = perc, fill = name_clean)) +
  geom_bar(stat = "identity") +
  #geom_hline(yintercept = c(0, .25, .5, .75, 1), lty = "dashed") +
  coord_polar(theta = "x") +
  scale_fill_manual(values = rev(vlce_colours)) +
  facet_wrap(~str_to_title(protected)) +
  theme_void() +
  theme(legend.position = "bottom") +
  labs(fill = NULL)

```

Figure 1: BEC proportions

```{r bec-prop}
bec <- vlce %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  group_by(protected, zone) %>%
  summarize(count = sum(n, na.rm = T)) %>%
  mutate(perc = count / sum(count))

bec %>%
  ggplot(aes(x = zone, y = perc, fill = str_to_title(protected))) +
  geom_col(position = "dodge") +
  labs(x = "Zone",
       y = "Percent",
       fill = NULL) +
  #coord_flip() +
  scale_y_continuous(labels = scales::percent) +
  theme(legend.position = c(0.8, 0.8))

# BEC subzone proportion of BC
bec_subzone_scatter <- vlce %>%
  group_by(protected, subzone) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count)) %>%
  dplyr::select(!count) %>%
  pivot_wider(names_from = protected, values_from = perc) %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  ggplot(aes(x = protected, y = unprotected, colour = zone)) +
  geom_point() +
  scale_y_continuous(labels = scales::percent) +
  scale_x_continuous(labels = scales::percent) +
  labs(colour = NULL,
       x = "Protected",
       y = "Unprotected") +
  geom_abline(slope = 1, intercept = 0, lty = "dashed") +
  scale_colour_igv()

ggsave(here::here("outputs", "bec_subzone_scatter.png"), 
       plot = bec_subzone_scatter, device = "png")

# Protected portions of subzones
per_prot_zone <- vlce %>% 
  group_by(subzone, protected) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count)) %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  filter(protected == "protected") %>%
  mutate(zone = fct_reorder(zone, perc)) 

n_zones <- per_prot_zone %>%
  count(zone)

bec_subzone_boxes <- per_prot_zone %>%
  left_join(n_zones) %>%
  mutate(label = paste0(zone, " (", n, ")"),
         label = fct_reorder(label, perc)) %>%
  ggplot() +
  geom_boxplot(aes(x = perc, y = label)) +
  geom_vline(aes(xintercept = 0.17, colour = "2020 Biodiversity Target (Global)"), lty = "dashed") +
  geom_vline(aes(xintercept = 0.10, colour = "2010 Biodiversity Target (Ecoregions)"), lty = "dashed") +
  scale_colour_aaas() +
  scale_x_continuous(labels = scales::percent) +
  labs(x = "% Protected",
       y = NULL,
       fill = NULL,
       colour = NULL) +
  theme(legend.position = c(0.75, 0.2))

ggsave(here::here("outputs", "bec_subzone_boxes.png"), 
       plot = bec_subzone_boxes, device = "png")

bec_prot_donut <- vlce %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  group_by(zone, protected) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count)) 

bec_perc_terr <- bec_prot_donut %>%
  summarize(count = sum(count)) %>%
  mutate(perc_terr = count / sum(count))

w <- sort(bec_perc_terr$perc_terr)
pos <- 0.5 * (cumsum(w) + cumsum(c(0, w[-length(w)])))

bec_conch_df <- bec_prot_donut %>%
  filter(protected == "protected") %>%
  left_join(keys$bec) %>%
  #left_join(bec_perc_terr, by = c("zone" = "zone")) %>%
  mutate(zone_nm = fct_reorder(zone_nm, perc),
         zone = fct_reorder(zone, perc, .fun = max)) 

write_csv(bec_conch_df, here::here("outputs", "bec_prop_protected.csv"))

bec_conch <- bec_conch_df %>%
  ggplot(aes(x = fct_reorder(zone, perc), y = perc, fill = zone_nm)) +
  geom_bar(stat = "identity") +
  geom_hline(aes(yintercept = 0.17, colour = "2020 Protection Target (Global)"), lty = "dotted", size = 1) +
  geom_hline(aes(yintercept = 0.25, colour = "2025 Protection Target (Canada)"), lty = "dotted", size = 1) +
  geom_hline(aes(yintercept = 0.154, colour = "2020 BC Proportion Protected"), lty = "dotted", size = 1) +
  geom_hline(aes(yintercept = 0.3), lty = "dotted", size = 1) +
  geom_text(aes(x = 1, y = 0.135), label = "15.4%") +
  geom_text(aes(x = 1, y = 0.19), label = "17%") +
  geom_text(aes(x = 1, y = 0.27), label = "25%", colour = "black") +
  geom_text(aes(x = 1, y = 0.32, label = "30%")) +
  scale_colour_aaas() +
  scale_fill_igv() +
  coord_polar() +
  theme_void() + 
  theme(axis.line = element_blank(),
        axis.title = element_blank(), 
        axis.ticks = element_blank(), 
        axis.text.y = element_blank(),
        #legend.position = "bottom",
        legend.box = "vertical") +
  labs(fill = NULL,
       colour = NULL)


ggsave(here::here("outputs", "bec_conch.png"), plot = bec_conch, device = "png",
       height = 5.5, width = 8.5)

bec_bar <- bec_conch_df %>%
  ggplot(aes(x = fct_reorder(zone, perc), y = perc, fill = zone_nm)) +
  geom_bar(stat = "identity") +
  geom_hline(yintercept = 0.17, colour = "#3B4992FF", lty = "dotted", size = 1) +
  geom_hline(yintercept = 0.10, colour = "#EE0000FF", lty = "dotted", size = 1) +
  geom_hline(yintercept = 0.154, colour = "#008B45FF", lty = "dotted", size = 1) +
  geom_text(aes(x = 3.2, y = .105), label = "2010 Ecoregional Protection Target (10%)", colour = "black") +
  geom_text(aes(x = 3, y = .175), label = "2020 Global Biodiversity Target (17%)") +
  geom_text(aes(x = 3.2, y = .159), label = "BC Overall Proportion Protected (15.4%)") +
  scale_colour_aaas() +
  scale_fill_igv() +
  labs(x = NULL,
       y = "% Protected",
       fill = NULL,
       colour = NULL) +
  scale_y_continuous(labels = scales::percent)

ggsave(here::here("outputs", "bec_bar.png"), plot = bec_bar, device = "png")

vlce %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  group_by(zone, protected) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count),
         zone = fct_reorder(zone, perc)) %>%
  filter(protected == "protected") %>%
  ggplot() +
  geom_point(aes(x = perc, y = zone)) +
  geom_vline(xintercept = c(.10, .17), colour = "red") +
  scale_x_continuous(lim = c(0, .302), labels = scales::percent) +
  labs(x = "%Protected",
       y = NULL,
       fill = NULL)

for_legend <- bec_conch_df %>% 
  ggplot(aes(x = zone, y = count, fill = zone_nm)) +
  geom_col() +
  scale_fill_igv() +
  labs(fill = NULL)


bec_legend <- cowplot::get_legend(for_legend + theme(legend.box.margin = margin(0, 0, 0, -500)))
```

VLCE Conch

```{r vlce-conch}
vlce_conch_df <- vlce %>%
  group_by(vlce, protected) %>%
  summarize(count = sum(n, na.rm = T)) %>%
  mutate(perc = count / sum(count)) %>%
  filter(protected == "protected") %>%
  left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
  dplyr::select(name_clean, protected, perc, count) %>%
  ungroup()

vlce_conch <- vlce_conch_df %>%
  ggplot(aes(x = fct_reorder(as.factor(vlce), perc), y = perc, fill = fct_reorder(name_clean, vlce, .fun = max))) +
  geom_bar(stat = "identity") +
  geom_hline(aes(yintercept = 0.17, colour = "2020 Protection Target (Global)"), lty = "dotted", size = 1) +
  geom_hline(aes(yintercept = 0.25, colour = "2025 Protection Target (Canada)"), lty = "dotted", size = 1) +
  geom_hline(aes(yintercept = 0.154, colour = "2020 BC Proportion Protected"), lty = "dotted", size = 1) +
  geom_hline(aes(yintercept = 0.33), lty = "dotted", size = 1) +
  geom_text(aes(x = 1, y = 0.135), label = "15.4%") +
  geom_text(aes(x = 1, y = 0.19), label = "17%") +
  geom_text(aes(x = 1, y = 0.27), label = "25%", colour = "black") +
  geom_text(aes(x = 1, y = 0.35, label = "33%")) +
  scale_colour_aaas() +
  scale_fill_manual(values = rev(vlce_colours)) +
  coord_polar() +
  theme_void() + 
  theme(axis.line = element_blank(),
        axis.title = element_blank(), 
        axis.ticks = element_blank(), 
        axis.text.y = element_blank(),
        #legend.position = "bottom",
        legend.box = "vertical") +
  labs(fill = NULL,
       colour = NULL) +
  guides(fill = guide_legend(order = 2),
         colour = guide_legend(order = 1))

vlce_conch
  
ggsave(here::here("outputs", "vlce_conch.png"), plot = vlce_conch, device = "png",
       height = 5.5, width = 8.5)
```



T-Test Plots

```{r t-tests}
t_tests_b <- t_tests %>%
  filter(!variable %in% c("elev", "slope")) %>%
  mutate(p.value = p.value * 496) ## bonferroni correction

t_tests_zonal <- t_tests_b %>%
  separate(subzone, into = c("zone", "subzone")) %>% 
  mutate(sig0.05 = ifelse(p.value < 0.05, T, F),
         sig0.01 = ifelse(p.value < 0.01, T, F)) %>%
  group_by(zone, variable) %>%
  summarize(sig0.05 = sum(sig0.05),
            sig0.01 = sum(sig0.01),
            total = n()) %>%
  mutate(per_sig0.05 = sig0.05 / total,
         per_sig0.01 = sig0.01 / total) %>%
  left_join(keys$continuous)

t_tests_b %>%
  separate(subzone, into = c("zone", "subzone")) %>% 
  mutate(sig0.05 = ifelse(p.value < 0.05, T, F),
         sig0.01 = ifelse(p.value < 0.01, T, F)) %>%
  group_by(variable) %>%
  summarize(sig0.05 = sum(sig0.05),
            sig0.01 = sum(sig0.01),
            total = n()) %>%
  mutate(per_sig0.05 = sig0.05 / total,
         per_sig0.01 = sig0.01 / total)

t_tests_scatter <- t_tests_zonal %>%
  filter(!variable %in% c("slope", "elev")) %>%
  mutate(axis_label = paste0(var_long, " (", unit, ")")) %>%
  ggplot(aes(x = per_sig0.01, y = axis_label)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(aes(colour = zone), size = 2, width = 0) +
  #geom_text_repel(aes(label = zone), max.overlaps = 20) +
  labs(x = "%of Significant (P < 0.01) Tests",
       y = NULL) +
  scale_colour_igv() +
  labs(colour = NULL) +
  scale_x_continuous(labels = scales::percent, limits = c(0.5, 1)) 

t_tests_scatter

ggsave(here::here("outputs", "t_tests_scatter.png"), plot = t_tests_scatter, device = "png",
       height = 5.5, width = 8.5)
  
```

```{r t-test-2}
t_tests_2 <- t_tests %>%
  separate(subzone, into = c("zone", "subzone")) %>% 
  filter(!variable %in% c("elev", "slope")) %>%
  mutate(#p.value = p.value + 0.0001,
         significant = case_when(p.value < 0.0001 / 496 ~ "< 0.001",
                                 p.value < 0.001 / 496 ~ "0.001",
                                 p.value < 0.01 / 496 ~ "0.01",
                                 p.value < 0.05 / 496 ~ "0.05",
                                 p.value < 0.10 / 496 ~ "0.10",
                                 TRUE ~ "Not Signficant")) %>%
  left_join(keys$continuous)

t_tests_2 %>%
  ggplot(aes(x = var_long, y = p.value, colour = significant)) +
  #geom_boxplot(outlier.alpha = 0) +
  geom_jitter(alpha = 0.3) +
  #geom_hline()
  #scale_y_continuous(trans = "log10") +
  labs(x = NULL,
       y = "P-value",
       colour = "Significance") +
  scale_colour_igv() +
  scale_fill_igv() +
  facet_wrap(~zone, scales = "free") +
  geom_hline(aes(yintercept = 0.01 / 496, colour = "0.01"), lty = "dashed") +
  geom_hline(aes(yintercept = 0.05 / 496, colour = "0.05"), lty = "dashed")

t_tests_2 %>%
  ggplot(aes(x = zone, y = p.value)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(aes(colour = significant), alpha = 0.5) +
  #geom_hline()
  #scale_y_continuous(trans = "log10") +
  labs(x = NULL,
       y = "P-value",
       colour = "Significance") +
  scale_colour_igv() +
  scale_fill_igv() +
  facet_wrap(~var_long, scales = "free") +
  geom_hline(aes(yintercept = 0.01 / 496, colour = "0.01"), lty = "dashed") +
  geom_hline(aes(yintercept = 0.05 / 496, colour = "0.05"), lty = "dashed") +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5))

t_tests_2 %>%
  ggplot(aes(x = var_long, y = p.value)) + 
  geom_jitter(aes(colour = significant)) +
  scale_y_continuous(limits = c(0, 0.05))

t_tests_2 %>%
  group_by(variable, significant) %>%
  summarize(count = n()) %>%
  mutate(perc = count / sum(count)) %>%
  left_join(keys$continuous) %>%
  ggplot(aes(x = var_long, y = perc, fill = significant)) +
  geom_col() +
  scale_fill_scico_d(palette = "batlow") +
  labs(x = NULL, 
       y = "Proportion Significant",
       fill = "Significance Threshold") +
  scale_y_continuous(labels = scales::percent)

t_tests_2 %>%
  ggplot(aes(x = var_long, y = p.value)) +
  geom_boxplot(outlier.alpha = 0)
```


Structure 3D Plot

```{r 3d-structure-subzone}
library(plot3D) # in use for 3d plot
library(plot3Drgl) # not in use
library(gridGraphics) # probably in use

bec_cols <- pal_igv("default")(16)
structure <- read_csv(here::here("data", "structure_means.csv"))

for_arrows_subzone_full <- structure %>%
  dplyr::select(!ends_with("total_biomass")) %>%
  dplyr::select(protected, subzone, starts_with("mean")) %>%
  pivot_wider(names_from = protected, values_from = starts_with("mean")) %>%
  #%>%
  drop_na() #%>%
  #filter(zone != "SBPS") %>%

for_arrows_subzone <- structure %>% 
  dplyr::select(!ends_with("total_biomass")) %>%
  dplyr::select(!starts_with("sd") & !count) %>%
  mutate(across(starts_with("mean"), scale)) %>%
  pivot_wider(names_from = protected, values_from = starts_with("mean")) %>%
  mutate(elev_cv_dif = mean_elev_cv_protected - mean_elev_cv_unprotected,
         loreys_height_dif = mean_loreys_height_protected - mean_loreys_height_unprotected,
         percentage_first_returns_above_2m_dif = mean_percentage_first_returns_above_2m_protected - mean_percentage_first_returns_above_2m_unprotected) %>%
  dplyr::select(subzone, ends_with("dif")) %>%
  mutate(vector_length = sqrt((elev_cv_dif)^2 + (loreys_height_dif)^2 + (percentage_first_returns_above_2m_dif)^2)) %>%
  filter(vector_length > 1) %>%
  left_join(for_arrows_subzone_full) %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  mutate(id = group_indices(., zone))

arrows_zones <- for_arrows_subzone %>%
  pull(zone) %>% 
  unique()
  
subzone_cols <- bec_cols[(keys$bec %>% pull(zone)) %in% arrows_zones]

for_arrows_subzone <- for_arrows_subzone %>% 
  mutate(label = paste(zone, subzone, sep = "-"),
         id = group_indices(., label))

# get the legend from this
for_arrows_subzone %>%
  ggplot(aes(x = id, y = vector_length, fill = label)) +
  geom_col() +
  labs(fill = "Zone-Subzone") +
  scale_fill_scico_d()

subzones_3d <- function() {
  #plotrgl()
  arrows3D(for_arrows_subzone$mean_elev_cv_protected, 
           for_arrows_subzone$mean_loreys_height_protected, 
           for_arrows_subzone$mean_percentage_first_returns_above_2m_protected,
           for_arrows_subzone$mean_elev_cv_unprotected, 
           for_arrows_subzone$mean_loreys_height_unprotected, 
           for_arrows_subzone$mean_percentage_first_returns_above_2m_unprotected,
           colvar = for_arrows_subzone$id,
           #col = subzone_cols,
           colkey = FALSE,
           lwd = 1, 
           d = 3, 
           #bty = "g",
           ticktype = "detailed",
           xlab = "\n Elevation Covariance",
           ylab = "\n Forest Height (m)",
           zlab = "\n Canopy Cover (%)",
           type = "cone"
           #xlim = c(-3, 3),
           #ylim = c(-3, 3),
           #zlim = c(-3, 3)
           )
  
  points3D(for_arrows_subzone$mean_elev_cv_protected, 
           for_arrows_subzone$mean_loreys_height_protected, 
           for_arrows_subzone$mean_percentage_first_returns_above_2m_protected,
           labels = for_arrows_subzone$label,
           colvar = for_arrows_subzone$id,
           #col = subzone_cols,
           add = T,
           colkey = F,
           type = "p",
           pch = 19,
           clab = "BEC Zone"#,
           #colkey = list(at = c(1:max(for_arrows_subzone$id)), side = 4,
           #             addlines = T, labels = unique(for_arrows_subzone$zone))
           )
}


subzones_3d_scatter <- plot_grid(subzones_3d, bec_legend,
          scale = c(1, 1))

subzones_3d_scatter

ggsave(here::here("outputs", "subzones_3d_scatter.png"),
       plot = subzones_3d_scatter, device = "png")

zones_img <- image_read(here::here("outputs", "subzones_3d_scatter.png")) 

zones_img %>% image_trim() %>% image_write(here::here("outputs", "subzones_3d_scatter.png"), format = "png")

```

```{r 3d-structure-zone}
# weighted means to account for varying subzone size

for_arrows_zone <- structure %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  dplyr::select(!starts_with("sd")) %>%
  group_by(protected, zone) %>%
  summarize(mean_elev_cv = weighted.mean(mean_elev_cv, count),
            mean_percentage_first_returns_above_2m = weighted.mean(mean_percentage_first_returns_above_2m, count),
            mean_loreys_height = weighted.mean(mean_loreys_height, count),
            mean_total_biomass = weighted.mean(mean_total_biomass, count)) %>%
  pivot_wider(names_from = protected, values_from = starts_with("mean")) %>%
  #filter(zone != "SBPS") %>%
  mutate(id = row_number())

structure %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  dplyr::select(!starts_with("sd")) %>%
  group_by(protected, zone) %>%
  summarize(mean_elev_cv = weighted.mean(mean_elev_cv, count),
            mean_percentage_first_returns_above_2m = weighted.mean(mean_percentage_first_returns_above_2m, count),
            mean_loreys_height = weighted.mean(mean_loreys_height, count),
            mean_total_biomass = weighted.mean(mean_total_biomass, count)) %>%
  mutate(across(starts_with("mean"), scale)) %>%
  pivot_wider(names_from = protected, values_from = starts_with("mean")) %>%
  mutate(elev_cv_dif = mean_elev_cv_protected - mean_elev_cv_unprotected,
         loreys_height_dif = mean_loreys_height_protected - mean_loreys_height_unprotected,
         percentage_first_returns_above_2m_dif = mean_percentage_first_returns_above_2m_protected - mean_percentage_first_returns_above_2m_unprotected) %>%
  dplyr::select(zone, ends_with("dif")) %>%
  mutate(vector_length = sqrt((elev_cv_dif)^2 + (loreys_height_dif)^2 + (percentage_first_returns_above_2m_dif)^2)) %>%
  arrange(desc(vector_length))
  

zones_3d <- function() {

  arrows3D(for_arrows_zone$mean_elev_cv_protected, 
           for_arrows_zone$mean_loreys_height_protected, 
           for_arrows_zone$mean_percentage_first_returns_above_2m_protected,
           for_arrows_zone$mean_elev_cv_unprotected, 
           for_arrows_zone$mean_loreys_height_unprotected, 
           for_arrows_zone$mean_percentage_first_returns_above_2m_unprotected,
           col = bec_cols,
           lwd = 1, 
           d = 3, 
           #bty = "g",
           ticktype = "detailed",
           xlab = "\n Elevation Covariance",
           ylab = "\n Forest Height (m)",
           zlab = "\n Canopy Cover (%)",
           type = "cone")
  
  points3D(for_arrows_zone$mean_elev_cv_protected, 
           for_arrows_zone$mean_loreys_height_protected, 
           for_arrows_zone$mean_percentage_first_returns_above_2m_protected,
           labels = for_arrows_zone$zone,
           colvar = for_arrows_zone$id,
           col = bec_cols, 
           add = T,
           colkey = F,
           type = "p",
           pch = 19,
           clab = "BEC Zone"
           #colkey = list(at = c(1:max(for_arrows_zone$id)), side = 4,
          #               addlines = T, labels = for_arrows_zone$zone)
           )
}

bec_legend <- cowplot::get_legend(for_legend + theme(legend.box.margin = margin(0, 0, 0, -150)))

zones_3d_scatter <- plot_grid(zones_3d) + bec_legend #& plot_annotation(theme = theme(plot.margin = margin(0, 5, 0, 5, "cm")))

ggsave(here::here("outputs", "zones_3d_scatter.png"),
       plot = zones_3d_scatter, device = "png")

# cropping because i couldn't figure it out in GGPLOT lol

zones_img <- image_read(here::here("outputs", "zones_3d_scatter.png")) 

zones_img %>% image_trim() %>% image_write(here::here("outputs", "zones_3d_scatter.png"), format = "png")

```

structure scatterplots

```{r structure-scatter}
structure_scatter <- structure %>% 
  dplyr::select(!c(starts_with("sd"), "count")) %>%
  pivot_longer(starts_with("mean"), 
               names_to = "variable", 
               values_to = "value") %>%
  pivot_wider(names_from = protected, values_from = value) %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  mutate(variable = str_remove(variable, "mean_")) %>%
  left_join(keys$continuous) %>%
  ggplot(aes(x = protected, y = unprotected, colour = zone)) +
  geom_point(alpha = 0.75) +
  geom_abline(intercept = 0, slope = 1, lty = "dashed", alpha = 0.5) +
  facet_wrap(~var_long, scales = "free") +
  scale_colour_igv() +
  labs(colour = NULL,
       x = "Protected",
       y = "Unprotected")

ggsave(here::here("outputs", "structure_scatter.png"),
       plot = structure_scatter, device = "png")
```

```{r structure-violin}
struct_cloud <- structure %>% 
  dplyr::select(!c(starts_with("sd"), "count")) %>%
  pivot_longer(starts_with("mean"), 
               names_to = "variable", 
               values_to = "value") %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  mutate(variable = str_remove(variable, "mean_")) %>%
  left_join(keys$continuous) %>%
  dplyr::select(protected, value, var_long, zone) %>%
  ggplot(aes(x = fct_rev(protected), y = value)) +
  ggdist::stat_halfeye(normalize = "panels",
                       ## custom bandwidth
                       adjust = .5, 
                       ## adjust height
                       width = .6, 
                       #alpha = 0.75,
                       ## move geom to the right
                       justification = -.2, 
                       ## remove slab interval
                       .width = 0, 
                       point_colour = NA, aes(fill = str_to_title(protected))) +
  geom_boxplot(width = 0.15,
               outlier.color = NA) +
  geom_point(size = 1.3,
             alpha = .3,
             position = position_jitter(
               seed = 1, width = 0.1
             )) +
  coord_flip() +
  facet_wrap(~var_long, scales = "free") +
  scale_fill_igv() +
  labs(x = NULL,
       y = NULL,
       fill = NULL,
       colour = NULL) +
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        legend.position = "bottom")

struct_cloud

ggsave(here::here("outputs", "structure_raincloud.png"), plot = struct_cloud, device = "png")
```

Latitude Disturbance

```{r lat-dist}
bc_bounds <- read_sf(here::here("data", "shapefiles", "bc_boundary.shp"))

lat_bbox <- bc_bounds %>% st_transform(4326) %>% st_bbox()

lat_range <- lat_bbox$ymax - lat_bbox$ymin

lat_agg <- 0.5
  
min_lat <- lat_dist %>%
  pull(latitude) %>%
  min()

max_lat <- lat_dist %>%
  pull(latitude) %>%
  max()
    
lat_df <- lat_dist %>% 
  group_by(protected, latitude, Change_Attribution) %>%
  summarize(count = sum(count)) %>%
  mutate(latitude = latitude - min_lat,
         per_lat = latitude / max_lat,
         lat_real = (per_lat * lat_range) + lat_bbox$ymin,
         lat_real = lat_real %/% lat_agg * lat_agg) %>%
  mutate(Change_Attribution = case_when(Change_Attribution == 1 ~ 1,
                                        Change_Attribution == 2 ~ 2,
                                        TRUE ~ 0)) %>%
  group_by(protected, lat_real, Change_Attribution) %>%
  summarize(count = sum(count)) %>%
  mutate(perc = count / sum(count)) %>%
  left_join(keys$disturbance, by = c("Change_Attribution" = "class_val"))

lat_plot <- lat_df %>%
  ggplot(aes(x = lat_real, y = perc, fill = class)) +
  geom_area() +
  facet_wrap(~str_to_title(protected)) +
  scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
  coord_flip(ylim = c(.75, 1)) +
  labs(x = expression("Latitude " ( degree~N)),
       y = "% Disturbed",
       fill = NULL) +
  theme(legend.position = "bottom") +
  scale_fill_manual(values = dist_colours) +
  scale_x_continuous(breaks = seq(48, 60, 3))

lat_prot <- lat_df %>%
  group_by(lat_real, protected) %>%
  summarize(count = sum(count)) %>%
  mutate(perc = count / sum(count)) %>%
  filter(protected == "protected") %>%
  ggplot(aes(x = lat_real, y = perc)) +
  #geom_smooth() +
  #geom_point() +
  geom_area() +
  scale_y_continuous(labels = scales::percent, limits = c(0, 1)) +
  labs(x = NULL,
       y = "% Protected",
       fill = NULL) +
  theme(axis.text.y = element_blank(),
        legend.position = c(0.7, 0.85)) +
  coord_flip() +
  scale_x_continuous(breaks = seq(48, 60, 3))
  

latitude_disturbance_plot <- plot_grid(lat_plot, lat_prot,
          rel_widths = c(3, 1),
          align = 'h', axis = 'tb')

ggsave(here::here("outputs", "latitude_disturbance_plot.png"),
       plot = latitude_disturbance_plot, device = "png",
       height = 5.5,
       width = 8.5)
```

```{r}
agg_vals = seq(from = 0.05, to = 1, by = 0.05)

output <- tibble()

for (val in agg_vals) {
  lat_df <- lat_dist %>% 
    group_by(protected, latitude, Change_Attribution) %>%
    summarize(count = sum(count)) %>%
    mutate(latitude = latitude - min_lat,
           per_lat = latitude / max_lat,
           lat_real = (per_lat * lat_range) + lat_bbox$ymin,
           lat_real = lat_real %/% val * val) %>%
    mutate(Change_Attribution = case_when(Change_Attribution == 1 ~ 1,
                                          Change_Attribution == 2 ~ 2,
                                          TRUE ~ 0)) %>%
    group_by(protected, lat_real, Change_Attribution) %>%
    summarize(count = sum(count)) %>%
    mutate(perc = count / sum(count)) %>%
    left_join(keys$disturbance, by = c("Change_Attribution" = "class_val"))
  
  lat_df["agg_val"] = val
  
  output <- bind_rows(output, lat_df)
}

output %>%
  ggplot(aes(x = lat_real, y = perc, fill = class)) +
    geom_area() +
    facet_grid(cols = vars(protected),
               rows = vars(agg_val)) +
    scale_y_continuous(labels = scales::percent_format(accuracy = 5L)) +
    coord_flip(ylim = c(.75, 1)) +
    labs(x = expression("Latitude " ( degree~N)),
         y = "% Disturbed",
         fill = NULL) +
    theme(legend.position = "bottom") +
    scale_fill_manual(values = dist_colours)

ggsave(here::here("outputs", "lol.png"), device = "png",
       width = 7.5,
       height = 49)
```

 

overall vlce for more t tests maybe?

```{r forest-wetland}
forest <- vlce %>%
  left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
  group_by(protected, subzone, forest) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count, na.rm = T)) %>%
  dplyr::select(!count) %>%
  filter(forest == "Forest") %>%
  pivot_wider(names_from = protected, values_from = perc) %>%
  drop_na() %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  rename(class = forest)

wetland <- vlce %>%
  left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
  group_by(protected, subzone, wetland) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count)) %>%
  dplyr::select(!count) %>%
  filter(wetland == "Wetland") %>%
  pivot_wider(names_from = protected, values_from = perc) %>%
  drop_na() %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  rename(class = wetland)

f_w <- bind_rows(forest, wetland)

f_w_scatter <- f_w %>%
  ggplot(aes(x = protected, y = unprotected, colour = zone)) +
  geom_point() +
  geom_abline(intercept = 0, slope = 1, lty = "dashed", alpha = 0.5) +
  scale_colour_igv() +
  scale_y_continuous(labels = scales::percent) +
  scale_x_continuous(labels = scales::percent) +
  #coord_cartesian(xlim = c(0, .15), ylim = c(0, .15)) +
  facet_wrap(~class) +
  labs(colour = NULL,
       x = "Protected",
       y = "Unprotected") +
  coord_equal()

ggsave(here::here("outputs", "forest_wetland_scatter.png"), plot = f_w_scatter,
       device = "png")

```


```{r map}
bec_sf <- read_sf(here::here("data", "shapefiles", "bec_zones.shp"))

bec_zones <- aggregate(x = bec_sf$are_sqm, by = list(bec_sf$zone),
                       FUN = sum, na.rm = TRUE,
                       do.union = T)

bec_zones = bec_sf %>% group_by(zone) %>%
  summarize(area = sum(are_sqm, na.rm = TRUE))

bc_bounds <- read_sf(here::here("data", "shapefiles", "bc_boundary.shp"))

bec_zones_terr <- bec_zones %>%
  st_intersection(bc_bounds)

parks <- read_sf(here::here("data", "shapefiles", "bc_ppa.shp"))

provinces <- getData(country = "Canada", level = 1)

prov_sf <- provinces %>%
  st_as_sf() %>%
  st_transform(3005)

bbox <- bc_bounds %>%
  st_bbox()

data("countriesHigh")

#rworldxtra

usa <- countriesHigh %>%
  st_as_sf() %>%
  st_transform(3005) %>%
  filter(ADMIN == "United States of America")

bec_zones_terr <- bec_zones_terr %>%
  left_join(keys$bec)

map <- ggplot() +
  geom_sf(data = bec_zones_terr, aes(fill = fct_reorder(zone_nm, zone)), colour = NA) +
  geom_sf(data = parks, aes(linetype = "Protected Areas"), fill = "Green") +
  #geom_sf(data = bc_bounds, fill = NA, size = 0.5) +
  geom_sf(data = usa) +
  geom_sf(data = prov_sf %>% filter(NAME_1 != "British Columbia"), fill = "grey50") +
  scale_linetype_manual(labels = "Protected Areas", values = "solid") +
  scale_fill_igv() +
  labs(fill = NULL,
       linetype = NULL) +
  coord_sf(xlim = c(bbox$xmin - 300000, bbox$xmax), 
           ylim = c(bbox$ymin, bbox$ymax),
           crs = 3005) +
  theme(panel.background = element_rect(fill = "lightblue"),
        legend.position = c(.15, .4),
        legend.background = element_rect(fill = NA),
        legend.box.background = element_rect(fill = "white"))

ggsave(here::here("outputs", "bec_map.png"), plot = map, device = "png", height = 7, width = 10)
```


```{r bec-table}
bec <- read_sf(here::here("data", "shapefiles", "bec_subzones.shp"))

bec_df <- bec %>% 
  mutate(area = st_area(.)) %>%
  st_drop_geometry()

num_subzones <- bec_df %>%
  group_by(zone, subzone) %>%
  summarize(area = sum(area)) %>%
  group_by(zone) %>%
  summarize(n_subzones = n())

bec_area <- bec_df %>%
  group_by(zone) %>%
  summarize(area = sum(area))

bec_ppa <- read_sf(here::here("data", "shapefiles", "bc_ppa_bec_hres.shp"))

bec_ppa_df <- bec_ppa %>%
  mutate(area = st_area(.),
         protected = ifelse(!is.na(IUCN_CA), T, F)) %>%
  st_drop_geometry()

bec_perc <- bec_ppa_df %>% 
  group_by(zone, protected) %>% 
  summarize(area = sum(area)) %>%
  mutate(perc = area / sum(area)) %>%
  filter(protected == T) %>%
  drop_na() %>%
  dplyr::select(zone, perc_protected = perc)

bec_table <- keys$bec %>% 
  left_join(num_subzones) %>%
  left_join(bec_area) %>%
  left_join(bec_perc) %>%
  mutate(area = as.integer(as.numeric(area / 10000)),
         perc_protected = paste0(round(as.numeric(perc_protected * 100), 1), "%"))

write_csv(bec_table, here::here("outputs", "bec_table.csv"))
```

```{r}
bec_table <- read_csv(here::here("outputs", "bec_table.csv"))

bec_table

bec_ft <- flextable(bec_table) %>%
  set_header_labels(bec_ft, 
                    zone = "Zone",
                    zone_nm = "Zone Name",
                    n_subzones = "# of Subzones",
                    area = "Area (ha)",
                    perc_protected = "# Protected") %>%
  set_caption("Number of subzones, total area, and percent protected by BEC Zone",
              autonum = run_autonum(seq_id = "tab", bkm = "bec-table"))
```


```{r prot-other}
perc_prot <- vlce %>%
  group_by(subzone, protected) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count)) %>%
  filter(protected == "protected") %>%
  dplyr::select(!c(count, protected)) %>%
  rename(perc_prot = perc)
  
perc_both <- vlce %>% 
  left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
  mutate(other = ifelse(forest == "Other", "Other", "Not-other")) %>%
  group_by(subzone, other) %>%
  summarize(count = sum(n)) %>% 
  mutate(perc_other = count / sum(count)) %>%
  filter(other == "Other") %>%
  dplyr::select(subzone, perc_other) %>%
  left_join(perc_prot) %>%
  separate(subzone, into = c("zone", "subzone")) 



perc_both <- perc_other %>%
  dplyr::select(subzone, perc_other = protected) 

summary(lm(perc_other ~ perc_prot, data = perc_both))$r.squared 

perc_both %>%
  ggplot(aes(x = perc_prot, y = perc_other, colour = zone)) +
  geom_point() +
  scale_colour_igv() +
  geom_abline(intercept = 0, slope = 1, lty = "dashed") +
  coord_equal(xlim = c(0, 1), ylim = c(0, 1)) +
  labs(x = "% Protected",
       y = "% Other",
       colour = "Zone") +
  scale_x_continuous(labels = scales::percent) +
  scale_y_continuous(labels = scales::percent)

perc_other %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  ggplot(aes(x = protected, y = unprotected, colour = zone)) +
  geom_point() +
  scale_colour_igv() +
  labs(x = "Protected",
       y = "Unprotected",
       colour = "Zone")


```

```{r subzones-missing-counterparts}
vlce %>% 
  left_join(keys$vlce, by = c("vlce" = "class_val")) %>%
  mutate(other = ifelse(forest == "Other", "Other", "Not-other")) %>%
  group_by(subzone, protected, other) %>%
  summarize(count = sum(n)) %>%
  mutate(perc = count / sum(count)) %>%
  dplyr::select(!count) %>%
  pivot_wider(names_from = protected, values_from = perc) %>%
  filter(other == "Other") %>%
  filter(is.nan(protected) | is.nan(unprotected))
```

```{r}
structure %>%
  dplyr::select(!starts_with("sd")) %>%
  pivot_longer(starts_with("mean")) %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  ggplot(aes(x = str_to_title(protected), y = value)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(aes(colour = zone), alpha = 0.5) +
  facet_wrap(~name, scales = "free") +
  scale_colour_igv() +
  labs(colour = "Zone",
       x = NULL,
       y = NULL)
```

```{r subzone-tables}
library(gt)

for_gt <- structure %>%
  separate(subzone, into = c("zone", "subzone")) %>%
  dplyr::select(!starts_with("sd")) %>%
  group_by(protected, zone) %>%
  summarize(mean_elev_cv = weighted.mean(mean_elev_cv, count),
            mean_percentage_first_returns_above_2m = weighted.mean(mean_percentage_first_returns_above_2m, count),
            mean_loreys_height = weighted.mean(mean_loreys_height, count),
            mean_total_biomass = weighted.mean(mean_total_biomass, count)) %>%
  dplyr::select(!mean_total_biomass) %>%
  pivot_wider(names_from = protected, values_from = starts_with("mean")) %>%
  mutate(mean_elev_change = (1 - (mean_elev_cv_protected / mean_elev_cv_unprotected)) * 100,
         mean_per_change = (1 - (mean_percentage_first_returns_above_2m_protected / mean_percentage_first_returns_above_2m_unprotected)) * 100,
         mean_lor_change = (1 - (mean_loreys_height_protected / mean_loreys_height_unprotected)) * 100) %>%
  select(zone,
         mean_elev_cv_protected,
         mean_elev_cv_unprotected,
         mean_elev_change,
         mean_percentage_first_returns_above_2m_protected,
         mean_percentage_first_returns_above_2m_unprotected,
         mean_per_change,
         mean_loreys_height_protected,
         mean_loreys_height_unprotected,
         mean_lor_change)
  

write_csv(for_gt, file = here::here("outputs", "vector_table.csv"))



for_gt %>%
  gt(rowname_col = "zone") %>%
  fmt_number(starts_with("mean")) %>%
  fmt_percent(ends_with("change")) %>%
  tab_spanner(
    label = "Elevation Covariance",
    columns = starts_with("mean_elev")
  ) %>%
  cols_label(
    mean_elev_cv_protected = "Protected",
    mean_elev_cv_unprotected = "Unprotected",
    mean_elev_change = "% Change"
  ) %>%
  tab_style(
    style = cell_fill(color = "grey80"),
    locations = cells_body(columns = c(mean_elev_cv_protected,
                                       mean_elev_cv_unprotected,
                                       mean_elev_change),
                           rows = (zone %in% highlighter$elev))
  ) %>%
  fmt_percent(columns = starts_with("mean_per")) %>%
  tab_spanner(
    label = "Canopy Cover (%)",
    columns = starts_with("mean_per")
  ) %>%
  cols_label(
    mean_percentage_first_returns_above_2m_protected = "Protected",
    mean_percentage_first_returns_above_2m_unprotected = "Unprotected",
    mean_per_change = "% Change"
  ) %>%
  tab_style(
    style = cell_fill(color = "grey80"),
    locations = cells_body(columns = c(mean_percentage_first_returns_above_2m_protected,
                                       mean_percentage_first_returns_above_2m_unprotected,
                                       mean_per_change),
                           rows = (zone %in% highlighter$lor))
  ) %>%
  tab_spanner(
    label = "Canopy Height (m)",
    columns = starts_with("mean_lor")
  ) %>%
  cols_label(
    mean_loreys_height_protected = "Protected",
    mean_loreys_height_unprotected = "Unprotected",
    mean_lor_change = "% Change"
  ) %>%
  tab_style(
    style = cell_fill(color = "grey80"),
    locations = cells_body(columns = c(mean_loreys_height_protected,
                                       mean_loreys_height_unprotected,
                                       mean_lor_change),
                           rows = (zone %in% highlighter$per))
  )
  

```

```{r}
zones <- for_gt %>%
  pull(zone)

highlighter <- for_gt %>%
  dplyr::select(zone, ends_with("change")) %>%
  pivot_longer(cols = ends_with("change")) %>%
  separate(name, into = c("mean", "var", "change")) %>%
  filter(abs(value) >= 5) %>%
  group_split(var) %>%
  set_names("elev", "lor", "per") %>%
  map(pull, zone)
  

index_highlight <- function(valid_zone) {
  index_out <- zones %in% valid_zone
  
  return(which(index_out))
}

highlight_indexes <- map(highlighter, index_highlight)

for_gt %>%
  mutate(across(ends_with("protected"), round, digits = 2),
         across(ends_with("change"), round, digits = 2),
         across(ends_with("change"), paste0, "%"),
         across(starts_with("mean_perc"), paste0, "%")) %>%
  flextable() %>%
  add_header_row(values = c("Zone", "Elevation Covariance", "Canopy Cover (%)", "Canopy Height (m)"),
                 colwidths = c(1, 3, 3, 3)) %>%
  set_header_labels(
    mean_elev_cv_protected = "Protected",
    mean_elev_cv_unprotected = "Unprotected",
    mean_elev_change = "% Change",
    mean_percentage_first_returns_above_2m_protected = "Protected",
    mean_percentage_first_returns_above_2m_unprotected = "Unprotected",
    mean_per_change = "% Change",
    mean_loreys_height_protected = "Protected",
    mean_loreys_height_unprotected = "Unprotected",
    mean_lor_change = "% Change",
    zone = "Zone") %>%
  theme_booktabs(bold_header = T) %>%
  align(part = "all", align = "center") %>%
  border_inner_v(border = fp_border(color = "grey"), part = "all") %>%
  merge_v(part = "header") %>%
  colformat_num() %>%
  autofit() %>%
  bg(i = highlight_indexes$elev,
     j = 2:4,
     bg = "grey") %>%
  bg(i = highlight_indexes$per,
     j = 5:7,
     bg = "grey") %>%
  bg(i = highlight_indexes$lor,
     j = 8:10,
     bg = "grey")
```
